+++ src/command/tools/fmtOutPutReport.ts
@@ -1,9 +1,9 @@
 import * as vscode from 'vscode';
 import { EFormatChannel } from '../../globalEnum';
-import { FormatCore } from '../../provider/Format/FormatProvider';
+import { FormatCore, FormatCoreWrap } from '../../provider/Format/FormatProvider';
 
 export async function fmtOutPutReport(document: vscode.TextDocument): Promise<void> {
-    const TextEdit: vscode.TextEdit[] = FormatCore({
+    const TextEdit: vscode.TextEdit[] = FormatCoreWrap(FormatCore({
         document,
         options: {
             tabSize: 4,
@@ -13,7 +13,7 @@ export async function fmtOutPutReport(document: vscode.TextDocument): Promise<vo
         fmtEnd: document.lineCount - 1,
         from: EFormatChannel.byFormatAllFile,
         needDiff: true,
-    });
+    }));
 
     const edit: vscode.WorkspaceEdit = new vscode.WorkspaceEdit();
     edit.set(document.uri, TextEdit);