+++ src/provider/Format/FormatProvider.ts
@@ -1,14 +1,12 @@
 /* eslint no-magic-numbers: ["error", { "ignore": [0,1,2,-999] }] */
 /* eslint-disable max-lines-per-function */
-import type * as vscode from 'vscode';
+import * as vscode from 'vscode';
 import { getFormatConfig } from '../../configUI';
 import type { TConfigs } from '../../configUI.data';
 import type { TAhkFileData } from '../../core/ProjectManager';
 import { pm } from '../../core/ProjectManager';
 import { EFormatChannel } from '../../globalEnum';
 import type { TBrackets } from '../../tools/Bracket';
-import type { TDiffMap } from './tools/fmtDiffInfo';
-import { fmtDiffInfo } from './tools/fmtDiffInfo';
 import { getMatrixFileBrackets } from './tools/getMatrixFileBrackets';
 import { getMatrixMultLine } from './tools/getMatrixMultLine';
 import { getMatrixTopLabe } from './tools/getMatrixTopLabe';
@@ -16,6 +14,8 @@ import { fn_Warn_thisLineText_WARN } from './TWarnUse';
 import type { TLnStatus } from './wantRefactor/getDeepKeywords';
 import { EFmtMagicStr, getDeepKeywords } from './wantRefactor/getDeepKeywords';
 
+import type { TFmtCore, TFmtCoreMap } from './FormatType';
+import { fmtDiffInfo } from './tools/fmtDiffInfo';
 import { getSwitchRange, inSwitchBlock } from './wantRefactor/SwitchCase';
 
 type TFmtCoreArgs = {
@@ -30,6 +30,36 @@ type TFmtCoreArgs = {
     needDiff: boolean,
 };
 
+export function FormatCoreWrap(map: TFmtCoreMap): vscode.TextEdit[] {
+    const textEditList: vscode.TextEdit[] = [];
+
+    // - if all files not changed
+    //
+    // | ms       | 0.0.23    | 0.0.24b | x   |
+    // | -------- | --------- | ------- | --- |
+    // | 88-files | 1600~1800 | 450~600 | 3X  |
+    // | 29-files | 700~800   | 50~80   | 10X |
+    //
+    // - if all files changed
+    //
+    // | ms       | 0.0.23    | 0.0.24b   | x  |
+    // | -------- | --------- | --------- | -- |
+    // | 88-files | 1600~1800 | 1600~1800 | 1X |
+    // | 29-files | 700~800   | 700~800   | 1X |
+    //
+    // test by FormatAllFile
+
+    for (const { line, oldText, newText } of map.values()) {
+        if (oldText !== newText) {
+            //      ^ performance optimization
+            const endCharacter: number = Math.max(newText.length, oldText.length);
+            const range = new vscode.Range(line, 0, line, endCharacter);
+            textEditList.push(new vscode.TextEdit(range, newText));
+        }
+    }
+    return textEditList;
+}
+
 export function FormatCore(
     {
         document,
@@ -38,14 +68,15 @@ export function FormatCore(
         fmtEnd,
         from,
     }: TFmtCoreArgs,
-): vscode.TextEdit[] {
+): TFmtCoreMap {
     const timeStart: number = Date.now();
 
+    const newFmtMap = new Map<number, TFmtCore>();
     /**
      * always update status
      */
     const AhkFileData: TAhkFileData | null = pm.updateDocDef(document);
-    if (AhkFileData === null) return [];
+    if (AhkFileData === null) return newFmtMap;
 
     const userConfigs: TConfigs['format'] = getFormatConfig();
     const {
@@ -53,7 +84,7 @@ export function FormatCore(
         useTopLabelIndent,
         AMasterSwitchUseFormatProvider,
     } = userConfigs;
-    if (!AMasterSwitchUseFormatProvider) return [];
+    if (!AMasterSwitchUseFormatProvider) return newFmtMap;
 
     const { DocStrMap, uri } = AhkFileData;
     const matrixTopLabe: readonly (0 | 1)[] = getMatrixTopLabe(AhkFileData, useTopLabelIndent);
@@ -68,9 +99,7 @@ export function FormatCore(
     // const memo: (Readonly<TLnStatus>)[] = [];
     // memo.push({ ...lnStatus });
     const switchRangeArray: vscode.Range[] = [];
-    const newTextList: vscode.TextEdit[] = [];
 
-    const DiffMap: TDiffMap = new Map();
     for (const AhkTokenLine of DocStrMap) {
         const { line, lStr } = AhkTokenLine;
         const lStrTrim: string = lStr.trim();
@@ -80,18 +109,20 @@ export function FormatCore(
             const occHotFix: number = status === EFmtMagicStr.caseA
                 ? occ + 1
                 : occ;
-            newTextList.push(fn_Warn_thisLineText_WARN({
-                DiffMap,
-                lStrTrim,
-                occ: occHotFix,
-                brackets: matrixBrackets[line],
-                options,
-                switchDeep: inSwitchBlock(lStrTrim, line, switchRangeArray),
-                topLabelDeep: matrixTopLabe[line],
-                MultLine: matrixMultLine[line],
-                formatTextReplace,
-                userConfigs,
-            }, AhkTokenLine));
+            newFmtMap.set(
+                line,
+                fn_Warn_thisLineText_WARN({
+                    lStrTrim,
+                    occ: occHotFix,
+                    brackets: matrixBrackets[line],
+                    options,
+                    switchDeep: inSwitchBlock(lStrTrim, line, switchRangeArray),
+                    topLabelDeep: matrixTopLabe[line],
+                    MultLine: matrixMultLine[line],
+                    formatTextReplace,
+                    userConfigs,
+                }, AhkTokenLine),
+            );
         } else if (line > fmtEnd) {
             break;
         }
@@ -109,10 +140,11 @@ export function FormatCore(
         // memo.push({ ...lnStatus });
     }
 
-    if (DiffMap.size > 0) {
+    // FIXME:
+    if (formatTextReplace) {
         const { fsPath } = uri;
         fmtDiffInfo({
-            DiffMap,
+            newFmtMap,
             fsPath,
             timeStart,
             from,
@@ -121,7 +153,7 @@ export function FormatCore(
 
     // console.log({ ms: Date.now() - timeStart, memo });
 
-    return newTextList;
+    return newFmtMap;
 }
 
 export const FormatProvider: vscode.DocumentFormattingEditProvider = {
@@ -130,13 +162,13 @@ export const FormatProvider: vscode.DocumentFormattingEditProvider = {
         options: vscode.FormattingOptions,
         _token: vscode.CancellationToken,
     ): vscode.ProviderResult<vscode.TextEdit[]> {
-        return FormatCore({
+        return FormatCoreWrap(FormatCore({
             document,
             options,
             fmtStart: 0,
             fmtEnd: document.lineCount - 1,
             from: EFormatChannel.byFormatAllFile,
             needDiff: true,
-        });
+        }));
     },
 };