+++ src/provider/Format/tools/fmtDiffInfo.ts
@@ -2,12 +2,10 @@ import * as path from 'node:path';
 import { EVersion } from '../../../Enum/EVersion';
 import type { EFormatChannel } from '../../../globalEnum';
 import { fmtLog } from '../../vscWindows/log';
-
-type TStrChange = [string, string];
-export type TDiffMap = Map<number, TStrChange>;
+import type { TFmtCore } from '../FormatType';
 
 type TDiffParm = {
-    DiffMap: TDiffMap,
+    newFmtMap: Map<number, TFmtCore>,
     fsPath: string,
     timeStart: number,
     from: EFormatChannel,
@@ -15,29 +13,34 @@ type TDiffParm = {
 
 export function fmtDiffInfo(
     {
-        DiffMap,
+        newFmtMap,
         fsPath,
         timeStart,
         from,
     }: TDiffParm,
 ): void {
-    const msg: string[] = [
-        '\n',
-        '-----------Format Diff Start--------------------------------',
-        `${from} ${EVersion.formatRange} "${path.basename(fsPath)}", ${Date.now() - timeStart} ms`,
-    ];
+    const msg: string[] = [];
 
-    for (const [line, [oldStr, newStr]] of DiffMap) {
-        msg.push(
-            `line : ${line}`,
-            oldStr,
-            newStr,
-        );
+    for (const [line, { hasOperatorFormat, oldText, newText }] of newFmtMap) {
+        if (hasOperatorFormat) {
+            msg.push(
+                `line : ${line}`,
+                oldText,
+                newText,
+            );
+        }
     }
 
-    msg.push('-----------Format Diff End----------------------------------');
+    if (msg.length > 0) {
+        fmtLog.info([
+            '\n',
+            '-----------Format Diff Start--------------------------------',
+            `${from} ${EVersion.formatRange} "${path.basename(fsPath)}", ${Date.now() - timeStart} ms`,
+            ...msg,
+            '-----------Format Diff End----------------------------------',
+        ].join('\n'));
+        // do not callDiff(diffVar);
+    }
 
-    fmtLog.info(msg.join('\n'));
-    // do not callDiff(diffVar);
     // using setTimeout call.
 }