+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 ## Next 0.0.24(2023-02-XX)
 
--fix: `1.e3` syntax-highlight
+- fix: `1.e3` syntax-highlight
+
+- perf: format 3X~10X
+  - if all files not changed
+
+  | ms       | 0.0.23    | 0.0.24  | x   |
+  | -------- | --------- | ------- | --- |
+  | 88-files | 1600~1800 | 450~600 | 3X  |
+  | 29-files | 700~800   | 50~80   | 10X |
+
+  - if all files changed
+
+  | ms       | 0.0.23    | 0.0.24    | x  |
+  | -------- | --------- | --------- | -- |
+  | 88-files | 1600~1800 | 1600~1800 | 1X |
+  | 29-files | 700~800   | 700~800   | 1X |
+
+  > use command `format All File`
 
 ## 0.0.23(2023-02-17)